drivers: sensor: tmp112: correct type
TMP112 registers are 16bit so there is no reason to use int64_t for
register values

+++ drivers/sensor/tmp112/tmp112.c
@@ -81,7 +81,7 @@ static int tmp112_attr_set(const struct device *dev,
 			   const struct sensor_value *val)
 {
 	const struct tmp112_config *cfg = dev->config;
-	int64_t value;
+	uint16_t value;
 	uint16_t cr;
 
 	if (chan != SENSOR_CHAN_AMBIENT_TEMP) {
